# wfe-restapi-client/src/main/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "runawfe api",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
    "/restapi/audit/process/{id}": {
      "post": {
        "tags": [
          "audit-controller"
        ],
        "summary": "getProcessLogs",
        "operationId": "getProcessLogsUsingPOST",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfeProcessLogFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WfeProcessLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/restapi/executor/user": {
      "get": {
        "tags": [
          "executor-controller"
        ],
        "summary": "getUser",
        "operationId": "getUserUsingGET",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeUser"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "executor-controller"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPUT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfeUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeUser"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "executor-controller"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPATCH",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfeUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/restapi/process/list": {
      "post": {
        "tags": [
          "process-controller"
        ],
        "summary": "getProcesses",
        "operationId": "getProcessesUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfePagedListFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfePagedListOfWfeProcess"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/sendSignal": {
      "post": {
        "tags": [
          "process-controller"
        ],
        "summary": "sendSignal",
        "operationId": "sendSignalUsingPOST",
        "parameters": [
          {
            "name": "ttlInSeconds",
            "in": "query",
            "description": "ttlInSeconds",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/signalReceiverIsActive": {
      "post": {
        "tags": [
          "process-controller"
        ],
        "summary": "signalReceiverIsActive",
        "operationId": "signalReceiverIsActiveUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/start": {
      "put": {
        "tags": [
          "process-controller"
        ],
        "summary": "startProcessByName",
        "operationId": "startProcessByNameUsingPUT",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeValidationExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/swimlane/{id}": {
      "put": {
        "tags": [
          "process-controller"
        ],
        "summary": "assignSwimlane",
        "operationId": "assignSwimlaneUsingPUT",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "executorId",
            "in": "query",
            "description": "executorId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "process-controller"
        ],
        "summary": "reassignSwimlaneById",
        "operationId": "reassignSwimlaneByIdUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/swimlanes": {
      "get": {
        "tags": [
          "process-controller"
        ],
        "summary": "getActiveProcessesSwimlanes",
        "operationId": "getActiveProcessesSwimlanesUsingGET",
        "parameters": [
          {
            "name": "namePattern",
            "in": "query",
            "description": "namePattern",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WfeSwimlane"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/upgrade": {
      "patch": {
        "tags": [
          "process-controller"
        ],
        "summary": "upgradeProcessesToDefinitionVersion",
        "operationId": "upgradeProcessesToDefinitionVersionUsingPATCH",
        "parameters": [
          {
            "name": "definitionId",
            "in": "query",
            "description": "definitionId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "version",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/{definitionId}/start": {
      "post": {
        "tags": [
          "process-controller"
        ],
        "summary": "startProcessById",
        "operationId": "startProcessByIdUsingPOST",
        "parameters": [
          {
            "name": "definitionId",
            "in": "path",
            "description": "definitionId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeValidationExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/{id}": {
      "get": {
        "tags": [
          "process-controller"
        ],
        "summary": "getProcess",
        "operationId": "getProcessUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeProcess"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "process-controller"
        ],
        "summary": "deleteProcess",
        "operationId": "deleteProcessUsingDELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/{id}/activate": {
      "patch": {
        "tags": [
          "process-controller"
        ],
        "summary": "activateProcess",
        "operationId": "activateProcessUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/process/{id}/cancel": {
      "patch": {
        "tags": [
          "process-controller"
        ],
        "summary": "cancelProcess",
        "operationId": "cancelProcessUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WfeProcessLogFilter": {
        "title": "WfeProcessLogFilter",
        "type": "object",
        "properties": {
          "createDateFrom": {
            "type": "integer",
            "format": "int64"
          },
          "createDateTo": {
            "type": "integer",
            "format": "int64"
          },
          "idFrom": {
            "type": "integer",
            "format": "int64"
          },
          "idTo": {
            "type": "integer",
            "format": "int64"
          },
          "includeSubprocessLogs": {
            "type": "boolean"
          },
          "nodeId": {
            "type": "string"
          },
          "tokenId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "ACTION",
              "ADMIN_ACTION",
              "ALL",
              "CREATE_TIMER",
              "NODE",
              "NODE_ENTER",
              "NODE_ERROR",
              "NODE_LEAVE",
              "PROCESS_ACTIVATE",
              "PROCESS_CANCEL",
              "PROCESS_END",
              "PROCESS_START",
              "PROCESS_SUSPEND",
              "RECEIVED_MESSAGE",
              "SEND_MESSAGE",
              "SUBPROCESS_END",
              "SUBPROCESS_START",
              "SWIMLANE_ASSIGN",
              "TASK",
              "TASK_ASSIGN",
              "TASK_CANCELLED",
              "TASK_CREATE",
              "TASK_DELEGATION",
              "TASK_END",
              "TASK_END_BY_ADMIN",
              "TASK_END_BY_SUBSTITUTOR",
              "TASK_ESCALATION",
              "TASK_EXPIRED",
              "TASK_REMOVED_ON_PROCESS_END",
              "TRANSITION",
              "VARIABLE",
              "VARIABLE_CREATE",
              "VARIABLE_DELETE",
              "VARIABLE_UPDATE"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}